Bug: Ensure rematch game has proper state reset
Prior to this commit, there was a bug where the server didn't properly
handle rematches. It would mean follow up games would not get a proper
game over, and pieces would spawn forever past the top of the board. The
fix is to properly reset the state for each player on a rematch